requirement.txt
django>=6.0.2
django-cors-headers>=4.9.0
django-filter>=25.2
django-import-export>=4.1.1
djangorestframework>=3.16.1
python-dotenv>=1.2.1
Pillow>=10.0.0
